refactor: move fetch & lib to /utils
