+++ lib/get.js
@@ -1,19 +1,8 @@
 module.exports = get
 
-var getProp = require('lodash.get')
-var setProp = require('lodash.set')
+var internals = module.exports.internals = {}
+internals.getProperties = require('../utils/get-properties')
 
-function get (state, basePath, path) {
-  var baseObject = getProp(state.session, basePath)
-  if (path === undefined) {
-    return baseObject
-  }
-
-  if (Array.isArray(path)) {
-    return path.reduce(function (properties, path) {
-      return setProp(properties, path, getProp(baseObject, path))
-    }, {})
-  }
-
-  return getProp(baseObject, path)
+function get (state, path) {
+  return internals.getProperties(state, 'account', path)
 }
