+++ lib/profile-fetch.js
@@ -4,5 +4,5 @@ var internals = module.exports.internals = {}
 internals.fetchProperties = require('../utils/fetch-properties')
 
 function fetch (state, path) {
-  return internals.fetchProperties(state, 'account', path)
+  return internals.fetchProperties(state, 'account.profile', path)
 }
