+++ package.json
@@ -49,6 +49,7 @@
     "npm-run-all": "^1.3.2",
     "rimraf": "^2.4.3",
     "semantic-release": "^6.0.3",
+    "simple-mock": "^0.5.0",
     "standard": "^5.2.2",
     "tap-spec": "^4.1.0",
     "tape": "^4.2.0",
