+++ tests/specs/index.js
@@ -14,4 +14,6 @@ require('./sign-out-test')
 require('./sign-up-test')
 require('./username-test')
 require('./utils-deserialise-test')
+require('./utils-get-properties-test')
+require('./utils-fetch-properties-test')
 require('./utils-serialise-test')
