+++ tests/specs/profile-fetch-test.js
@@ -1,50 +1,20 @@
+var simple = require('simple-mock')
 var test = require('tape')
-var nock = require('nock')
 
-var fetch = require('../../lib/fetch')
+var fetch = require('../../lib/profile-fetch')
+var internals = fetch.internals
 
-var baseURL = 'http://localhost:3000'
-var fetchResponse = require('../fixtures/fetch-profile.json')
-var state = {
-  url: baseURL,
-  session: {
-    id: 'sessionId123'
-  }
-}
+test('profileFetch', function (t) {
+  simple.mock(internals, 'fetchProperties').returnWith('foo')
+  var result = fetch('state', 'path')
 
-test('fetch profile details', function (t) {
-  t.plan(2)
+  t.deepEqual(internals.fetchProperties.lastCall.args, [
+    'state',
+    'account.profile',
+    'path'
+  ], 'calls utils.fetchProperties with "account.profile" basepath')
+  t.is(result, 'foo', 'returns result of fetchProperties')
 
-  var expectedObject = fetchResponse.data.attributes
-  expectedObject.id = fetchResponse.data.id
-
-  nock(baseURL)
-    .get('/session/account/profile')
-    .reply(200, JSON.stringify(fetchResponse))
-
-  fetch(state, 'account.profile')
-
-  .then(function (accountInfo) {
-    t.is(typeof accountInfo, 'object', 'returns account object')
-    t.deepEqual(accountInfo, expectedObject, 'contains correct info')
-  })
-
-  .catch(t.fail)
-})
-
-test('fetch profile fullName', function (t) {
-  t.plan(2)
-
-  nock(baseURL)
-    .get('/session/account/profile')
-    .reply(200, JSON.stringify(fetchResponse))
-
-  fetch(state, 'account.profile', 'fullName')
-
-  .then(function (fullName) {
-    t.is(typeof fullName, 'string', 'returns string')
-    t.is(fullName, fetchResponse.data.attributes.fullName, 'contains correct fullName')
-  })
-
-  .catch(t.fail)
+  simple.restore()
+  t.end()
 })
